+++ services/document-manager/backend/app/tasks/__init__.py
@@ -1,0 +1,1 @@
+from app.tasks import reorganize, transfer, verify
