+++ services/document-manager/backend/app/tasks/celery.py
@@ -1,0 +1,2 @@
+from app import make_celery
+celery = make_celery()
