+++ services/document-manager/backend/init.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 
 cd /app
-celery worker -A app.utils.tasks --detach --loglevel=info --logfile=/var/log/celery/celery.log --pidfile=/var/run/celery/celery.pid --concurrency=1
+celery worker -A app.tasks.celery --detach --loglevel=info --logfile=/var/log/celery/celery.log --pidfile=/var/run/celery/celery.pid --concurrency=1
 flask run
+# celery worker -A app.tasks.celery --loglevel=info --concurrency=1
